adapt install.sh for npm 2
npm 2 no longer has an engines field in its package.json, so don't check
for that. This also incorporates a bunch of cleanups suggested by
shellcheck, and drops the polite fiction that running Node < 0.8 is
going to give you an acceptable experience in 2014 with npm. It makes up
for this by adding 0.8 to the list of versions that gets the latest npm
by default.